--- a/Cargo.toml
+++ b/Cargo.toml
@@ -10,9 +10,17 @@ authors = ["Georg Wiese <georgwiese@gmail.com>"]
 [dependencies]
 serde = {version = "1.0.144", features = ["derive"]}
 serde-big-array = "0.4.1"
+rayon = "1.5.3"
 rmp-serde = "1.1.0"
 serde_json = "1.0.85"
 clap = {version = "3.2.22", features = ["derive"]}
 hmac-sha256 = "1.1.4"
 rand = "0.8.5"
 rand_chacha = "0.3.1"
+
+[[bench]]
+name = "bench"
+harness = false
+
+[dev-dependencies]
+criterion = "0.4.0"