--- a/Cargo.toml
+++ b/Cargo.toml
@@ -11,7 +11,6 @@ authors = ["Georg Wiese <georgwiese@gmail.com>"]
 serde = {version = "1.0.144", features = ["derive"]}
 serde-big-array = "0.4.1"
 rayon = "1.5.3"
-itertools = "0.10.5"
 rmp-serde = "1.1.0"
 serde_json = "1.0.85"
 clap = {version = "3.2.22", features = ["derive"]}