--- a/benches/bench.rs
+++ b/benches/bench.rs
@@ -2,43 +2,45 @@ use criterion::{criterion_group, criterion_main, Bencher, Criterion};
 
 use hash_based_signatures::signature::stateless_merkle::StatelessMerkleSignatureScheme;
 use hash_based_signatures::signature::winternitz::domination_free_function::D;
-use hash_based_signatures::signature::SignatureScheme;
+use hash_based_signatures::signature::{HashType, SignatureScheme};
 use rand::prelude::*;
 
+fn get_random_256bits() -> HashType {
+    let mut result = [0u8; 32];
+    let mut rng = thread_rng();
+    rng.fill_bytes(&mut result);
+    result
+}
+
+fn make_signature_scheme() -> StatelessMerkleSignatureScheme {
+    let seed = get_random_256bits();
+    StatelessMerkleSignatureScheme::new(seed, 16, 32, D::new(15))
+}
+
 fn key_generation(b: &mut Bencher) {
-    b.iter(|| {
-        let mut seed = [0u8; 32];
-        let mut rng = rand::thread_rng();
-        rng.fill_bytes(&mut seed);
-        StatelessMerkleSignatureScheme::new(seed, 16, 32, D::new(255));
-    })
+    b.iter(|| make_signature_scheme())
 }
 
 fn signing(b: &mut Bencher) {
-    let mut signature_scheme = StatelessMerkleSignatureScheme::new([0u8; 32], 16, 32, D::new(255));
+    let mut signature_scheme = make_signature_scheme();
     b.iter(|| {
-        // Sign a random message
-        let mut hash = [0u8; 32];
-        let mut rng = thread_rng();
-        rng.fill_bytes(&mut hash);
-        signature_scheme.sign(hash);
+        let msg = get_random_256bits();
+        signature_scheme.sign(msg);
     })
 }
 
 fn verification(b: &mut Bencher) {
-    let mut signature_scheme = StatelessMerkleSignatureScheme::new([0u8; 32], 16, 32, D::new(255));
-    let mut hash = [0u8; 32];
-    let mut rng = rand::thread_rng();
-    rng.fill_bytes(&mut hash);
-    let signature = signature_scheme.sign(hash);
+    let mut signature_scheme = make_signature_scheme();
+    let msg = get_random_256bits();
+    let signature = signature_scheme.sign(msg);
     b.iter(|| {
-        StatelessMerkleSignatureScheme::verify(signature_scheme.public_key(), hash, &signature)
+        StatelessMerkleSignatureScheme::verify(signature_scheme.public_key(), msg, &signature)
     })
 }
 
 fn criterion_benchmark(c: &mut Criterion) {
     let mut group = c.benchmark_group("benches");
-    group.sample_size(10);
+    group.sample_size(20);
 
     group.bench_function("key_generation", key_generation);
     group.bench_function("signing", signing);