--- a/example/readme.md
+++ b/example/readme.md
@@ -3,7 +3,7 @@
 This very readme will be signed by the following public key:
 
 ```
-702d39ca33cab5590ada460e4bc0d6821468cfd40ea593140c0e3002fd3c0412
+cef7b96b7fc47850cb01991c58c29bbfef733eefc6fd3f22e2d9b2bbd147a4e3
 ```
 
 You can verify the signature by running:
@@ -11,5 +11,5 @@ You can verify the signature by running:
 $ cargo run -- verify \
   example/readme.md \
   example/readme.md.signature \
-  702d39ca33cab5590ada460e4bc0d6821468cfd40ea593140c0e3002fd3c0412
+  cef7b96b7fc47850cb01991c58c29bbfef733eefc6fd3f22e2d9b2bbd147a4e3
 ```