… Cargo.lock
… example/readme.md.signature