Benchmark before (for `q: 16, depth: 32, d: 255`):
```
benches/key_generation  time:   [157.74 ms 161.27 ms 163.20 ms]
benches/signing         time:   [5.3891 s 5.4353 s 5.4958 s]
benches/verification    time:   [172.63 ms 176.43 ms 180.92 ms]
```

Benchmark after:
```
benches/key_generation  time:   [90.354 ms 93.308 ms 97.959 ms]
                        change: [-43.359% -41.661% -39.644%] (p = 0.00 < 0.05)
benches/signing         time:   [2.9921 s 3.0106 s 3.0282 s]
                        change: [-45.311% -44.610% -44.047%] (p = 0.00 < 0.05)
benches/verification    time:   [99.285 ms 117.08 ms 135.12 ms]
                        change: [-44.212% -38.430% -30.525%] (p = 0.00 < 0.05)

```